+++ frontend/src/admin/pages/[groupId]/workbook/[id]/index.vue
@@ -1,19 +1,7 @@
 <script setup lang="ts">
 import Button from '@/common/components/Atom/Button.vue'
-import InputItem from '@/common/components/Atom/InputItem.vue'
-import Modal from '@/common/components/Molecule/Modal.vue'
-import Switch from '@/common/components/Molecule/Switch.vue'
 import PaginationTable from '@/common/components/Organism/PaginationTable.vue'
-import TextEditor from '@/common/components/Organism/TextEditor.vue'
-import { NDatePicker } from 'naive-ui'
-import { ref } from 'vue'
 import IconTrash from '~icons/fa/trash-o'
-
-const show = ref(false)
-const title = ref('')
-const dateRange = ref<[number, number]>([1183135260000, Date.now()])
-const problemVisibleStat = ref(false)
-const difficultyVisibleStat = ref(false)
 </script>
 
 <template>
@@ -31,7 +19,7 @@ const difficultyVisibleStat = ref(false)
         <div>Total Problem: 10, Total Score: 100</div>
       </div>
       <div class="flex">
-        <Button class="mr-8 h-8" @click="show = true">+ Create</Button>
+        <Button class="mr-8 h-8">+ Create</Button>
         <Button class="h-8">Import</Button>
       </div>
     </div>
@@ -162,56 +150,6 @@ const difficultyVisibleStat = ref(false)
       </template>
     </PaginationTable>
   </div>
-  <Modal v-model="show" class="mx-8 h-[96%] w-full">
-    <div class="overflow-auto px-40">
-      <div class="flex flex-col">
-        <div class="mt-10">
-          <h1 class="text-gray-dark mr-6 inline text-2xl font-semibold">
-            Create Workbook
-          </h1>
-        </div>
-        <div class="bg-gray h-[1px]"></div>
-        <div class="mt-10 flex">
-          <div class="w-20 text-lg font-bold">Group</div>
-          <div class="flex items-center">NPC 중급반</div>
-        </div>
-        <div class="mt-6 flex">
-          <div class="w-20 text-lg font-bold">Title</div>
-          <InputItem v-model="title" required placeholder="Title" />
-        </div>
-        <h2 class="mt-8 text-lg">Description</h2>
-        <TextEditor class="h-[360px] w-full" />
-
-        <div class="mt-10">
-          <NDatePicker
-            v-model:value="dateRange"
-            type="datetimerange"
-            clearable
-          />
-        </div>
-        <div class="mt-10 flex">
-          <Switch v-model="problemVisibleStat" class="mr-8" label="Visible" />
-          <Switch
-            v-model="difficultyVisibleStat"
-            label="Problem Difficulty Visible"
-          />
-        </div>
-        <div class="bg-gray mt-10 h-[1px]"></div>
-        <div class="mt-8 flex justify-end">
-          <div class="flex">
-            <Button
-              class="mr-8 px-4 py-1"
-              color="gray-dark"
-              @click="show = false"
-            >
-              Cancel
-            </Button>
-            <Button class="px-4 py-1">Next</Button>
-          </div>
-        </div>
-      </div>
-    </div>
-  </Modal>
 </template>
 
 <route lang="yaml">
