+++ frontend/src/admin/pages/[groupId]/workbook/index.vue
@@ -1,7 +1,11 @@
 <script setup lang="ts">
 import Button from '@/common/components/Atom/Button.vue'
+import InputItem from '@/common/components/Atom/InputItem.vue'
+import Modal from '@/common/components/Molecule/Modal.vue'
 import Switch from '@/common/components/Molecule/Switch.vue'
 import PaginationTable from '@/common/components/Organism/PaginationTable.vue'
+import TextEditor from '@/common/components/Organism/TextEditor.vue'
+import { NDatePicker } from 'naive-ui'
 import { ref } from 'vue'
 import IconTrash from '~icons/fa/trash-o'
 
@@ -9,6 +13,12 @@ const props = defineProps<{
   groupId: string
 }>()
 const onoff = ref(true)
+
+const show = ref(false)
+const title = ref('')
+const dateRange = ref<[number, number]>([1183135260000, Date.now()])
+const problemVisibleStat = ref(false)
+const difficultyVisibleStat = ref(false)
 </script>
 
 <template>
@@ -19,7 +29,7 @@ const onoff = ref(true)
       <h1 class="text-gray-dark mr-6 inline text-2xl font-semibold">
         Workbook List
       </h1>
-      <Button>+ Create</Button>
+      <Button @click="show = true">+ Create</Button>
     </div>
     <PaginationTable
       :fields="[
@@ -83,6 +93,56 @@ const onoff = ref(true)
       </template>
     </PaginationTable>
   </div>
+  <Modal v-model="show" class="mx-8 h-[96%] w-full">
+    <div class="overflow-auto px-40">
+      <div class="flex flex-col">
+        <div class="mt-10">
+          <h1 class="text-gray-dark mr-6 inline text-2xl font-semibold">
+            Create Workbook
+          </h1>
+        </div>
+        <div class="bg-gray mt-2 h-[1px]"></div>
+        <div class="mt-10 flex">
+          <div class="w-20 text-lg font-bold">Group</div>
+          <div class="flex items-center">NPC 중급반</div>
+        </div>
+        <div class="mt-6 flex">
+          <div class="w-20 text-lg font-bold">Title</div>
+          <InputItem v-model="title" required placeholder="Title" />
+        </div>
+        <h2 class="mt-8 text-lg font-bold">Description</h2>
+        <TextEditor class="h-[360px] w-full" />
+        <h2 class="mt-8 text-lg font-bold">Period</h2>
+        <div>
+          <NDatePicker
+            v-model:value="dateRange"
+            type="datetimerange"
+            clearable
+          />
+        </div>
+        <div class="mt-10 flex">
+          <Switch v-model="problemVisibleStat" class="mr-8" label="Visible" />
+          <Switch
+            v-model="difficultyVisibleStat"
+            label="Problem Difficulty Visible"
+          />
+        </div>
+        <div class="bg-gray mt-10 h-[1px]"></div>
+        <div class="mt-8 flex justify-end">
+          <div class="flex">
+            <Button
+              class="mr-8 px-4 py-1"
+              color="gray-dark"
+              @click="show = false"
+            >
+              Cancel
+            </Button>
+            <Button class="px-4 py-1">Next</Button>
+          </div>
+        </div>
+      </div>
+    </div>
+  </Modal>
 </template>
 
 <route lang="yaml">
